add script to stop mac from auto-adjusting volume
I don't remember why I needed this, but there was a period where some
headphones where messing with an app, and it was screwing up my volume
settings. This fixed that.

In case you were wondering, this is why I fucking hate mac. I never had
to do anything as dumb as this on linux.